fix: Product Module separated from Photo
