--- src/modules/product/dto/photo/create-photo.dto.ts
@@ -1,7 +1,0 @@
-import { IsInstance, IsNotEmpty } from 'class-validator';
-
-export class CreatePhotoDto {
-  @IsNotEmpty()
-  @IsInstance(Buffer)
-  image: Buffer;
-}
