--- src/modules/product/dto/photo/update-photo.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreatePhotoDto } from './create-photo.dto';
-
-export class UpdatePhotoDto extends PartialType(CreatePhotoDto) {}
