--- src/modules/product/entities/photo.entity.ts
@@ -1,10 +1,0 @@
-import { Column, Entity, ObjectId, ObjectIdColumn } from 'typeorm';
-
-@Entity()
-export class Photo {
-  @ObjectIdColumn()
-  _id: ObjectId;
-
-  @Column({ type: 'bytea' })
-  image: Buffer;
-}
