+++ src/modules/product/mocks/productProvider.ts
@@ -2,9 +2,9 @@ import { getRepositoryToken } from '@nestjs/typeorm';
 import { Product } from '../entities/product.entity';
 import { ObjectLiteral, Repository } from 'typeorm';
 import { shopCategoryMock } from '../../../modules/shop/shop-category/mocks/shopCategoryProvider';
-import { CreateProductDto } from '../dto/product/create-product.dto';
-import { UpdateProductDto } from '../dto/product/update-product.dto';
-import { photoMock } from './photoProvider';
+import { CreateProductDto } from '../dto/create-product.dto';
+import { UpdateProductDto } from '../dto/update-product.dto';
+import { productPhotoMock as photoMock } from '../../photo/mocks/photoProvider';
 
 type MockRepository<T extends ObjectLiteral = any> = {
   [P in keyof Repository<T>]?: jest.Mock<any, any>;
