+++ src/modules/product/product.controller.ts
@@ -13,8 +13,8 @@ import {
   ParseFilePipeBuilder,
 } from '@nestjs/common';
 import { ProductService } from './product.service';
-import { CreateProductDto } from './dto/product/create-product.dto';
-import { UpdateProductDto } from './dto/product/update-product.dto';
+import { CreateProductDto } from './dto/create-product.dto';
+import { UpdateProductDto } from './dto/update-product.dto';
 import { ShopService } from '../shop/shop/shop.service';
 import { CategoryService } from '../category/category.service';
 import { ShopCategoryService } from '../shop/shop-category/shop-category.service';
