+++ src/modules/product/product.service.ts
@@ -1,23 +1,23 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
-import { CreateProductDto } from './dto/product/create-product.dto';
-import { UpdateProductDto } from './dto/product/update-product.dto';
+import { CreateProductDto } from './dto/create-product.dto';
+import { UpdateProductDto } from './dto/update-product.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Product } from './entities/product.entity';
-import { MongoRepository, Repository } from 'typeorm';
+import { Repository } from 'typeorm';
 import { ShopCategory } from '../shop/shop-category/entities/shop-category.entity';
-import { Photo } from './entities/photo.entity';
-import { CreatePhotoDto } from './dto/photo/create-photo.dto';
-import { UpdatePhotoDto } from './dto/photo/update-photo.dto';
-import { ObjectId } from 'mongodb';
+import { CreatePhotoDto } from '../photo/dto/create-photo.dto';
+import { UpdatePhotoDto } from '../photo/dto/update-photo.dto';
+import { PhotoService, PhotoType } from '../photo/photo.service';
 
 @Injectable()
 export class ProductService {
+  private type: PhotoType = 'product';
+
   constructor(
     @InjectRepository(Product)
     private readonly productRepository: Repository<Product>,
 
-    @InjectRepository(Photo, 'mongodb')
-    private readonly photoRepository: MongoRepository<Photo>
+    private readonly photoService: PhotoService
   ) {}
 
   async create(
@@ -25,9 +25,7 @@ export class ProductService {
     createProductDto: CreateProductDto,
     shopCategory: ShopCategory
   ) {
-    const photo = await this.photoRepository.save(
-      this.photoRepository.create(createPhotoDto)
-    );
+    const photo = await this.photoService.create(createPhotoDto, this.type);
 
     return await this.productRepository.save(
       this.productRepository.create({
@@ -45,17 +43,12 @@ export class ProductService {
       return products;
     }
 
-    const productPhotIds = products.map(({ photo }) =>
-      photo ? new ObjectId(photo) : undefined
-    );
-
-    const photos = await this.photoRepository.find({
-      _id: { $in: productPhotIds },
-    });
+    const productPhotIds = products.map(({ photo }) => photo);
+    const photos = await this.photoService.findAll(productPhotIds, this.type);
 
     return products.map((product, index) => ({
       ...product,
-      photo: productPhotIds[index] ? photos.shift()?.image : undefined,
+      photo: photos[index],
     }));
   }
 
@@ -68,17 +61,12 @@ export class ProductService {
       return products;
     }
 
-    const productPhotIds = products.map(({ photo }) =>
-      photo ? new ObjectId(photo) : undefined
-    );
-
-    const photos = await this.photoRepository.find({
-      _id: { $in: productPhotIds },
-    });
+    const productPhotIds = products.map(({ photo }) => photo);
+    const photos = await this.photoService.findAll(productPhotIds, this.type);
 
     return products.map((product, index) => ({
       ...product,
-      photo: productPhotIds[index] ? photos.shift()?.image : undefined,
+      photo: photos[index],
     }));
   }
 
@@ -89,11 +77,9 @@ export class ProductService {
       return product;
     }
 
-    const photo = await this.photoRepository.findOneBy({
-      _id: new ObjectId(product.photo),
-    });
+    const photo = await this.photoService.findOne(product.photo, this.type);
 
-    return { ...product, photo: photo?.image };
+    return { ...product, photo };
   }
 
   async update(
@@ -107,20 +93,16 @@ export class ProductService {
       throw new BadRequestException('This product does not exist');
     }
 
-    const existingPhoto = await this.photoRepository.findOneBy({
-      _id: new ObjectId(existingProduct.photo),
-    });
-
-    const updatedPhoto = {
-      ...existingPhoto,
-      ...updatePhotoDto,
-    };
-
-    await this.photoRepository.save(updatedPhoto);
+    const updatedPhoto = await this.photoService.update(
+      existingProduct.photo,
+      updatePhotoDto,
+      this.type
+    );
 
     const updatedProduct = {
       ...existingProduct,
       ...updateProductDto,
+      photo: updatedPhoto._id?.toString(),
     };
 
     return await this.productRepository.save(updatedProduct);
